+++ kujira/__init__.py
@@ -3,8 +3,8 @@
 import eventlet
 from flask import Flask
 from flask_socketio import SocketIO
-from kujira.blueprints import SERVER_BP, OSD_BP, POOL_BP, MON_BP, CLUSTER_BP
-from kujira.rest.controllers import osds, pools, servers, clusters, mons
+from kujira.blueprints import SERVER_BP, OSD_BP, POOL_BP, MON_BP, CLUSTER_BP, DISK_BP
+from kujira.rest.controllers import osds, pools, servers, clusters, mons, disks
 
 eventlet.monkey_patch()
 
@@ -20,6 +20,7 @@ def create_app():
     app.register_blueprint(POOL_BP)
     app.register_blueprint(MON_BP)
     app.register_blueprint(CLUSTER_BP)
+    app.register_blueprint(DISK_BP)
 
     app.config.from_object('config')
