+++ kujira/blueprints.py
@@ -7,3 +7,4 @@ POOL_BP = Blueprint('pool', __name__, url_prefix="/kujira/api/v1/pools")
 SERVER_BP = Blueprint('server', __name__, url_prefix="/kujira/api/v1/servers")
 MON_BP = Blueprint('mon', __name__, url_prefix="/kujira/api/v1/mons")
 CLUSTER_BP = Blueprint('cluster', __name__, url_prefix="/kujira/api/v1/clusters")
+DISK_BP = Blueprint('disk', __name__, url_prefix="/kujira/api/v1/disks")
